requirements.txt
fastapi
uvicorn[standard]
pytesseract
pdf2image
PyPDF2
Pillow
transformers
sentence-transformers
chromadb
python-dotenv
poppler-utils
httpx
python-multipart
